+++ src/ai_cv_agent/utils/__init__.py
@@ -1,0 +1,17 @@
+"""Utility functions and helpers for AI CV Agent."""
+
+from ai_cv_agent.utils.html_builder import generate_cv_html, ResumeData
+from ai_cv_agent.utils.job_fetcher import read_job_description
+from ai_cv_agent.utils.pdf_converter import html_to_pdf, html_to_pdf_async
+from ai_cv_agent.utils.profile_manager import read_user_profile
+from ai_cv_agent.utils.resume_mapper import convert_raw_resume_to_resume_data
+
+__all__ = [
+    "generate_cv_html",
+    "ResumeData", 
+    "read_job_description",
+    "html_to_pdf",
+    "html_to_pdf_async",
+    "read_user_profile",
+    "convert_raw_resume_to_resume_data",
+]
