+++ src/ai_cv_agent/utils/html_builder.py
