+++ src/ai_cv_agent/utils/job_fetcher.py
